Complete random partitioning
Our strategy for random partitioning is to open {numnodes} files, and
then iterate through all vectors in the dataset, randomly choosing one
of our files to write to. This _should_ mean that numpy doesn't require
the entire database to be loaded into Python memory (though we have not
tested this or confirmed it explicitly).

We also keep track of a `max_vector_count` for each node file, and if we
hit that max, we remove it from the list of files we can write vectors
to when iterating through the database. This means that all nodes should
have equal number of vectors, with a possible delta of 1 if the total
vectors / number of nodes is not an integer.

Co-authored-by: MithiJ <mithi_jethwa@brown.edu>