Adds get_sift script
Downloads from webpage via FTP, and then assigns attributes to each of
the vectors according to the method outlined in the ACORN paper (see
readme for more information).

The current format of predicates is a CSV file that is named the same as
the fvecs file, where each row represents one vector, and each column
represents an attribute.