Fix bug in #40
This was a C++ lifetime issue. The way that the FFI wrappers had been
constructed meant that the `IndexACORNFlat` destructor was being called,
which was freeing the `storage` memory. Thus the fix was just to call
the methods directly on the `UniquePtr` rather than dereferencing the
value in the wrapper.

Co-authored-by: Nick Thompson <ncthom91@gmail.com>