Kick build issue on unhappy path down the line
As described in the comment, the FFI integration doesn't seem to want to
play nice when compiling with code that throws errors from C++. I tried
just setting the `-frtti` flag explicitly in the build, but that didn't
seem to work. As this unhappy path is not mission-critical, I've just
added a TODO inline in the relevant test.