Update build script so that external FAISS lib is not required
It seems that up until this point, only the IndexACORN.cpp file has
actually been compiled by my toolchain. The rest of the `faiss` methods
were being provided by the dylib that was installed in the Dockefile.

This commit ensures that all FAISS is compiled from scratch. There is a
bit of an issue now with build times, as I haven't worked out how to
accurately notify the C++ compiler not to re-run every time.