Run query sets for SIFT (#41)
* Wrap in a dropshot server

* Rm tracing in favor of slog ecosystem for logging

Dropshot uses slog loggers

* Recreate CLI in bench.rs

* Fix tests

* Setup benchmarking harness

* Fix bug in #40

This was a C++ lifetime issue. The way that the FFI wrappers had been
constructed meant that the `IndexACORNFlat` destructor was being called,
which was freeing the `storage` memory. Thus the fix was just to call
the methods directly on the `UniquePtr` rather than dereferencing the
value in the wrapper.

Co-authored-by: Nick Thompson <ncthom91@gmail.com>

* Implement basic equality predicate for single attribute metadata

Co-authored-by: Oli Clive-Griffin <o.clivegriffin@gmail.com>

* Remove unnecessary &mut in search

* Add `new` method for PredicateQuery

* Add Result across the FFI boundary

* Kick build issue on unhappy path down the line

As described in the comment, the FFI integration doesn't seem to want to
play nice when compiling with code that throws errors from C++. I tried
just setting the `-frtti` flag explicitly in the build, but that didn't
seem to work. As this unhappy path is not mission-critical, I've just
added a TODO inline in the relevant test.

* Pass references in dataset trait and fvecs impl

* Template basic benchmark for predicate queries

* Update example.rs to be more descriptive

* Remove compiler warnings

* Comment out mmap.lock()

I don't think this is doing what I thought it was; and the data is
brought into memory when the `FlattenedVecs` representation is built in
any case.

* Update build script so that external FAISS lib is not required

It seems that up until this point, only the IndexACORN.cpp file has
actually been compiled by my toolchain. The rest of the `faiss` methods
were being provided by the dylib that was installed in the Dockefile.

This commit ensures that all FAISS is compiled from scratch. There is a
bit of an issue now with build times, as I haven't worked out how to
accurately notify the C++ compiler not to re-run every time.

* Attempt to speed up compilation

* Skeleton for benchmarking

---------

Co-authored-by: Nick Thompson <ncthom91@gmail.com>
Co-authored-by: Oli Clive-Griffin <o.clivegriffin@gmail.com>